requirements.txt
python-okx
Crypto
web3
eth_account